Delete device_pci.c
