+++ teeio-validator/tools/lside.c
@@ -20,8 +20,6 @@
 bool g_pci_log = false;
 int g_top_id = 0;
 int g_config_id = 0;
-int m_dev_fp = 0;
-uint32_t g_doe_extended_offset = 0;
 IDE_OPERATION g_ide_operation = IDE_OPERATION_LIST;
 IDE_TEST_CONFIG lside_test_config = {0};
 DEVCIES_CONTEXT devices_context = {0};
